+++ packages/eval/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ziro-agent/eval
 
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ziro-agent/agent@0.15.0
+
 ## 0.2.11
 
 ### Patch Changes